requirements/common.txt
Django==1.4.3
South==0.7.6
distribute==0.6.31
dj-database-url==0.2.1
django-taggit==0.10a1
psycopg2==2.4.6
wsgiref==0.1.2
